Added a fix for conversion of mods with options
Added a fix to BCML's convertor when trying to convert mods with options, due to there not being an `info.json` inside. Also modified the warning messages a little bit so that it's evident that the file comes from an option (by showing the option's folder name). The handling of warnings was also modified to account for some being passed as lists instead of individual strings.